requirements.txt
python-telegram-bot[job-queue]>=22
python-dotenv>=1.0
httpx>=0.28